Separate the Linux futex implementation
