+++ lib/io/event.rb
@@ -8,7 +8,3 @@ require_relative "event/support"
 require_relative "event/selector"
 require_relative "event/timers"
 require_relative "event/native"
-
-if defined?(IO::Event::Futex)
-	require_relative "event/futex"
-end
