+++ test/io/event/futex.rb
@@ -11,6 +11,28 @@ return unless defined?(IO::Event::Futex)
 describe IO::Event::Futex do
 	let(:buffer) {IO::Buffer.new(8)}
 	let(:futex) {subject.new(buffer)}
+	let(:uring_selector) do
+		unless defined?(IO::Event::Selector::URing)
+			skip "io_uring is not available"
+		end
+		
+		selector = IO::Event::Selector::URing.new(Fiber.current)
+		unless selector.respond_to?(:futex_wait)
+			selector.close
+			skip "io_uring futex operations are not available"
+		end
+		
+		selector
+	end
+	let(:waitv_selector) do
+		selector = uring_selector
+		unless selector.respond_to?(:futex_waitv)
+			selector.close
+			skip "io_uring futex waitv operations are not available"
+		end
+		
+		selector
+	end
 	
 	with "#value" do
 		it "stores and loads the value atomically" do
@@ -70,7 +92,7 @@ describe IO::Event::Futex do
 		end
 		
 		it "waits asynchronously for a signal" do
-			selector = IO::Event::Selector::URing.new(Fiber.current)
+			selector = uring_selector
 			result = nil
 			
 			fiber = Fiber.new do
@@ -94,7 +116,7 @@ describe IO::Event::Futex do
 		end
 		
 		it "uses the current scheduler" do
-			selector = IO::Event::Selector::URing.new(Fiber.current)
+			selector = uring_selector
 			scheduler = IO::Event::TestScheduler.new(selector: selector)
 			result = nil
 			
@@ -117,7 +139,7 @@ describe IO::Event::Futex do
 		end
 		
 		it "does not wait when the value has changed" do
-			selector = IO::Event::Selector::URing.new(Fiber.current)
+			selector = uring_selector
 			futex.value = 1
 			result = nil
 			
@@ -133,107 +155,112 @@ describe IO::Event::Futex do
 		end
 	end
 	
-	with ".wait_any" do
-		it "waits without blocking other Ruby threads when no scheduler is installed" do
-			first = subject.new(buffer, offset: 0)
-			second = subject.new(buffer, offset: 4)
-			
-			thread = Thread.new do
-				sleep 0.01
-				second.signal
+	if IO::Event::Futex.respond_to?(:wait_any)
+		with ".wait_any" do
+			it "exposes the maximum number of wait entries" do
+				expect(subject::WAITV_LIMIT).to be == 128
 			end
 			
-			expect(subject.wait_any([[first, 0], [second, 0]])).to be == 1
-		ensure
-			thread&.join
-		end
-		
-		it "does not wait without a scheduler when a value has changed" do
-			first = subject.new(buffer, offset: 0)
-			second = subject.new(buffer, offset: 4)
-			second.value = 1
-			
-			expect(subject.wait_any([[first, 0], [second, 0]])).to be_nil
-		end
-		
-		it "waits asynchronously for any futex to be signalled" do
-			selector = IO::Event::Selector::URing.new(Fiber.current)
-			unless selector.respond_to?(:futex_waitv)
-				skip "io_uring futex waitv operations are not available"
+			it "rejects more than the maximum number of wait entries" do
+				entries = Array.new(subject::WAITV_LIMIT + 1){[futex, 0]}
+				
+				expect do
+					subject.wait_any(entries)
+				end.to raise_exception(ArgumentError)
 			end
 			
-			first = subject.new(buffer, offset: 0)
-			second = subject.new(buffer, offset: 4)
-			result = nil
-			
-			fiber = Fiber.new do
-				result = selector.futex_waitv(Fiber.current, [[first, 0], [second, 0]])
-			end
-			fiber.transfer
-			
-			thread = Thread.new do
-				sleep 0.01
-				second.signal
-			end
-			
-			selector.select(1)
-			thread.join
-			
-			expect(result).to be == 1
-		ensure
-			selector&.close
-			thread&.join
-		end
-		
-		it "uses the current scheduler" do
-			selector = IO::Event::Selector::URing.new(Fiber.current)
-			unless selector.respond_to?(:futex_waitv)
-				skip "io_uring futex waitv operations are not available"
+			it "waits without blocking other Ruby threads when no scheduler is installed" do
+				first = subject.new(buffer, offset: 0)
+				second = subject.new(buffer, offset: 4)
+				
+				thread = Thread.new do
+					sleep 0.01
+					second.signal
+				end
+				
+				expect(subject.wait_any([[first, 0], [second, 0]])).to be == 1
+			ensure
+				thread&.join
 			end
 			
-			scheduler = IO::Event::TestScheduler.new(selector: selector)
-			first = subject.new(buffer, offset: 0)
-			second = subject.new(buffer, offset: 4)
-			result = nil
-			
-			Fiber.set_scheduler(scheduler)
-			Fiber.schedule do
-				result = subject.wait_any([[first, 0], [second, 0]])
+			it "does not wait without a scheduler when a value has changed" do
+				first = subject.new(buffer, offset: 0)
+				second = subject.new(buffer, offset: 4)
+				second.value = 1
+				
+				expect(subject.wait_any([[first, 0], [second, 0]])).to be_nil
 			end
 			
-			thread = Thread.new do
-				sleep 0.01
-				second.signal
+			it "waits asynchronously for any futex to be signalled" do
+				selector = waitv_selector
+				
+				first = subject.new(buffer, offset: 0)
+				second = subject.new(buffer, offset: 4)
+				result = nil
+				
+				fiber = Fiber.new do
+					result = selector.futex_waitv(Fiber.current, [[first, 0], [second, 0]])
+				end
+				fiber.transfer
+				
+				thread = Thread.new do
+					sleep 0.01
+					second.signal
+				end
+				
+				selector.select(1)
+				thread.join
+				
+				expect(result).to be == 1
+			ensure
+				selector&.close
+				thread&.join
 			end
 			
-			scheduler.run
-			
-			expect(result).to be == 1
-		ensure
-			Fiber.set_scheduler(nil)
-			thread&.join
-		end
-		
-		it "returns nil when a value has changed" do
-			selector = IO::Event::Selector::URing.new(Fiber.current)
-			unless selector.respond_to?(:futex_waitv)
-				skip "io_uring futex waitv operations are not available"
+			it "uses the current scheduler" do
+				selector = waitv_selector
+				
+				scheduler = IO::Event::TestScheduler.new(selector: selector)
+				first = subject.new(buffer, offset: 0)
+				second = subject.new(buffer, offset: 4)
+				result = nil
+				
+				Fiber.set_scheduler(scheduler)
+				Fiber.schedule do
+					result = subject.wait_any([[first, 0], [second, 0]])
+				end
+				
+				thread = Thread.new do
+					sleep 0.01
+					second.signal
+				end
+				
+				scheduler.run
+				
+				expect(result).to be == 1
+			ensure
+				Fiber.set_scheduler(nil)
+				thread&.join
 			end
 			
-			first = subject.new(buffer, offset: 0)
-			second = subject.new(buffer, offset: 4)
-			second.value = 1
-			result = :waiting
-			
-			fiber = Fiber.new do
-				result = selector.futex_waitv(Fiber.current, [[first, 0], [second, 0]])
+			it "returns nil when a value has changed" do
+				selector = waitv_selector
+				
+				first = subject.new(buffer, offset: 0)
+				second = subject.new(buffer, offset: 4)
+				second.value = 1
+				result = :waiting
+				
+				fiber = Fiber.new do
+					result = selector.futex_waitv(Fiber.current, [[first, 0], [second, 0]])
+				end
+				fiber.transfer
+				selector.select(1)
+				
+				expect(result).to be_nil
+			ensure
+				selector&.close
 			end
-			fiber.transfer
-			selector.select(1)
-			
-			expect(result).to be_nil
-		ensure
-			selector&.close
 		end
 	end
 end
